+++ src/components/ShopItem.tsx
@@ -4,7 +4,7 @@ import Images from '../assets/images';
 
 const ShopItem: React.FC = () => {
 	const cards = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
-	const cardItems = cards.map((item) => <ItemCard item={item} />);
+	const cardItems = cards.map((item) => <ItemCard key={item} item={item} />);
 	return <Layout className="shop-items">{cardItems}</Layout>;
 };
