+++ src/style/pages/_shop.scss
@@ -3,6 +3,7 @@
 	@include verticalPadding(2rem, 2rem);
 	flex-direction: row;
 	flex-wrap: wrap;
+	justify-content: space-between;
 	background: #ffffff;
 	.ant-card {
 		border: 0;
